feature(title-change): change title automatically on route change
Closes #32

Updated Root App Component, subscribing to NavigationEnd event and
changing Title through MetaService when route finishes changing.

Currently only works in Browser, as we're not supplying the entire Html
document. We'll have to look into fixing this later.

Routes now can contain a new property data : { title : 'whatever you
want' }, and it will automatically update the Title of the browser with
this