Feature(Angular 5.0): Upgrade to working copy of 5.0 (#538)
This works both in Development & Production. 🎁
(Apologies on the delay)

Now dotnet publish won't fail, and you can successfully build Production builds.

In the future, boot.server.PRODUCTION.ts probably won't be needed.

We'll be working on upgrading this to use the CLI version that Steve Sanderson has been working on the past few months - this will clean-up a lot of the configuration so that it's all handled by the CLI itself. (Aka: no more webpack files for us to deal with here) ✨

This should close out many issues:

#535
#482
(others as well, can't find them at the moment)