upgraded to angular-router-loader for proper angular 6 code splitting and lazy loading support
The prior approach with the older angular2-router-loader does not work anymore. We had to upgrade to the recommended angular-router-loader (see https://www.npmjs.com/package/angular-router-loader) and modify this webpack file accordingly. Works much better for us. In production, @ngtools/webpack  does not work for us, and the proposed changes have been verified in prod and are stable.